Optimization and Calibration of both modules
In the Simon Module, timing was adjusted and some user quality of life
changes were made.
The Alarm Module was not keeping time well at all. Time counter was
adjusted to allow much more accurate and consistent time keeping.
Any if statement that was in a loop that was called often was changed to
a switch statement for speed.